src/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.0
requests==2.32.5
seaborn==0.13.2
python-dotenv==1.2.1
psycopg==3.3.2
joblib==1.5.3
shap==0.50.0
langchain-core==1.2.7
torch==2.10.0
transformers==4.57.6
huggingface-hub==0.36.0
pillow==12.0.0
chromadb==1.4.1
tqdm==4.67.1
langchain-community==0.4.1
langchain==1.2.7
ipython==9.9.0
torchvision==0.25.0
mpmath==1.3.0
ipywidgets==8.1.8
scipy==1.16.3
jupyterlab_pygments==0.3.0
pip==24.3.1
attrs==25.4.0
distro==1.9.0
protobuf==3.20.3
tornado==6.5.4
Jinja2==3.1.6
nbformat==5.10.4
sympy==1.14.0
httpx==0.28.1
pydantic==2.12.5
ollama==0.6.1
langchain_ollama==1.0.0
langchain_chroma==1.1.0
sentencepiece==0.2.1
python-multipart==0.0.19